#16: extend data type range of timestamp to prevent overflow

+++ adma_ros2_driver/src/adma_driver.cpp
@@ -123,7 +123,7 @@ void ADMADriver::parseData(std::array<char, 856> recv_buf)
   //offset between UNIX and GNSS (in ms)
   unsigned long offset_gps_unix = 315964800000;
   uint32_t week_to_msec = 604800000;
-  unsigned long timestamp;
+  unsigned long long timestamp;
 
   // read Adma msg from UDP data packet
   if (protocol_version_ == "v3.2" || protocol_version_ == "v3.3.3") {
